requirements.txt
torch>=1.4.0
torchaudio>=0.5.0
git+https://github.com/PetrochukM/PyTorch-NLP.git
editdistance
pandas
PyYAML
matplotlib
tqdm